Application/Common/Models : create Dto for CU expenses -> CreateExpenseDto.cs : Description, Amount, Date, CategoryId; UpdateExpenseDto.cs : Id, Description, Amount, Date, CategoryId; ExpenseController.cs : CreateExpense endpoint receiving CreateExpenseDto, UpdateExpense endpoint receiving UpdateExpenseDto, DeleteResponse endpoint -> checking expense of this user and delete;

+++ ExpenseTracker.Web/Controllers/ExpenseController.cs
@@ -1,4 +1,5 @@
-﻿using ExpenseTracker.Application.Services.Interfaces;
+﻿using ExpenseTracker.Application.Common.Models;
+using ExpenseTracker.Application.Services.Interfaces;
 using ExpenseTracker.Domain.Entities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -40,21 +41,32 @@ namespace ExpenseTracker.Web.Controllers
         }
 
         [HttpPost]
-        public IActionResult CreateExpense([FromBody] Expense expense)
+        public IActionResult CreateExpense([FromBody] CreateExpenseDto model)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            var expense = new Expense()
+            {
+                Amount = model.Amount,
+                CategoryId = model.CategoryId,
+                Date = model.Date,
+                Description = model.Description,
+                UserId = userId
+            };
+
             _expenseService.CreateExpense(expense);
             return CreatedAtAction(nameof(GetExpenseById), new { id = expense.Id }, expense);
         }
 
         [HttpPut]
-        public IActionResult UpdateExpense(int id, [FromBody] Expense expense)
+        public IActionResult UpdateExpense(int id, [FromBody] UpdateExpenseDto model)
         {
-            if (id != expense.Id)
+            if (id != model.Id)
             {
                 return BadRequest("Id and Expense id must be equal!");
             }
@@ -64,7 +76,19 @@ namespace ExpenseTracker.Web.Controllers
                 return BadRequest(ModelState);
             }
 
-            _expenseService.UpdateExpense(expense);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            var expense = new Expense()
+            {
+                Id = model.Id,
+                Amount = model.Amount,
+                CategoryId = model.CategoryId,
+                Date = model.Date,
+                Description = model.Description,
+                UserId = userId
+            };
+
+            _expenseService.UpdateExpense(expense); 
 
             return NoContent(); 
         }
@@ -73,6 +97,11 @@ namespace ExpenseTracker.Web.Controllers
         public IActionResult DeleteExpense(int id)
         {
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var expense = _expenseService.GetExpenseById(id, userId);
+            if (expense == null)
+            {
+                return NotFound();
+            }
             _expenseService.DeleteExpense(id, userId);
             return NoContent(); 
         }
